// src/shared/api/generated/types/DocumentVersionDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type DocumentVersionDto = {
  /**
   * @type number
   */
  id: number;
  /**
   * @type number
   */
  versionNumber: number;
  /**
   * @type string
   */
  originalFilename: string;
  /**
   * @type string
   */
  extension: string;
  /**
   * @type string
   */
  mimeType: string;
  /**
   * @type string, int64
   */
  sizeBytes: string;
  /**
   * @minLength 64
   * @maxLength 64
   * @type string
   */
  sha256: string;
  /**
   * @type string, date-time
   */
  createdAt: string;
  /**
   * @type boolean
   */
  isCurrent: boolean;
};